macos-requirements.txt
PyObjC
AppKit
Foundations
python-dateutil
